Spec to illustrate error thrown when including fast_jsonapi
When trying to use the fast_jsonapi gem, we started getting an exception in unrelated `ActiveRecord` code like:

```
 Failure/Error:
   def #{name}_id
     association(:#{name}).reader.try(:id)

 NoMethodError:
   undefined method `has_scope?' for nil:NilClass
 # ./vendor/bundle/ruby/2.4.0/gems/activerecord-5.1.6/lib/active_record/reflection.rb:872:in `has_scope?'
 # ./vendor/bundle/ruby/2.4.0/gems/activerecord-5.1.6/lib/active_record/reflection.rb:873:in `has_scope?'
 # ./vendor/bundle/ruby/2.4.0/gems/activerecord-5.1.6/lib/active_record/associations/association.rb:288:in `skip_statement_cache?'
 # ./vendor/bundle/ruby/2.4.0/gems/activerecord-5.1.6/lib/active_record/associations/singular_association.rb:39:in `find_target'
 # ./vendor/bundle/ruby/2.4.0/gems/activerecord-5.1.6/lib/active_record/associations/association.rb:157:in `load_target'
 # ./vendor/bundle/ruby/2.4.0/gems/activerecord-5.1.6/lib/active_record/associations/association.rb:53:in `reload'
 # ./vendor/bundle/ruby/2.4.0/gems/activerecord-5.1.6/lib/active_record/associations/singular_association.rb:7:in `reader'
 # ./lib/extensions/has_one.rb:15:in `city_id'

```

This error was only showing when we had required in the `fast_jsonapi` and was happening in code that was not doing JSON serialization.

The reason for the error was that we were making a call to an `ActiveRecord` `select` where we were explicitly selecting the ID value for a `has_one` that was coming through another relation.

The `HasOne` monkey patch in `lib/extensions/has_one.rb` is the cause of this error.

The spec above illustrates the problem. If you run it, you will see the above exception. If you comment out the code in `lib/extensions/has_one.rb`, you will see the new spec passes.